Исправил получение данных с сервера
