+++ js/main.js
@@ -9,7 +9,6 @@ import {
 } from './form.js';
 
 import { addMap } from './map.js';
-import { getData } from './api.js';
 // Функции активации страницы
 const activatePage = () => {
   activateMapFilters();
@@ -22,6 +21,4 @@ const deactivatePage = () => {
 };
 
 deactivatePage();
-getData(addMap, activatePage);
-
-
+addMap(activatePage);
