+++ js/map.js
@@ -3,6 +3,7 @@ import {
   addressInput
 } from './form.js';
 import { createCard } from './card.js';
+import { getData } from './api.js';
 
 const START_COORDINATES = {
   lat: 35.681700,
@@ -28,29 +29,6 @@ const mainMarker = L.marker(
     icon: mainPin,
   },
 );
-const ALERT_SHOW_TIME = 5000;
-const ERROR_MASSAGE = 'Не удалось загрузить данные объявлений с сервера';
-
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 1000;
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = '150px';
-  alertContainer.style.top = '200px';
-  alertContainer.style.right = '150px';
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '30px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-};
 
 const addMarkers = (location, card) => {
   const adMarkerGroup = L.layerGroup().addTo(map);
@@ -73,7 +51,17 @@ const addMarkers = (location, card) => {
     );
 };
 
-const addMap = (onLoadCallback, data) => {
+const renderAdsOnMap = (data) => {
+  if (data) {
+    data.forEach((ad) => {
+      const location = ad.location;
+      const card = createCard(ad);
+      addMarkers(location, card);
+    });
+  }
+};
+
+const addMap = (onLoadCallback) => {
   map.on('load', () => {
     onLoadCallback();
   })
@@ -95,15 +83,7 @@ const addMap = (onLoadCallback, data) => {
       setAddressValue(address);
     });
   }
-  if (data) {
-    data.forEach((ad) => {
-      const location = ad.location;
-      const card = createCard(ad);
-      addMarkers(location, card);
-    });
-  } else {
-    showAlert(ERROR_MASSAGE);
-  }
+  getData(renderAdsOnMap);
 };
 
 const resetMap = () => {
